Fix jumping issues when having unequal size items
When the items are of different size and take some time to render, the virtual
scroller may be in a situation where there are items on top of the ones that
are displayed that have not been rendered and thus measured by the cache yet.
With variable sizes, this can be a problem, as we rely on the cache, or on the
default size to measure the correct padding and avoid jumping issues when
scrolling top.
This commit tries to address this issue but still needs further work.